Construct an initial pyproject.toml; enable "uv"
This provides metadata for the project, and the server's dependencies.

With the .toml, the project can now use "uv" for virtual environments
to run the server.

There is a bit of leakage for test dependencies, for an upcoming PR,
but it should not affect the pubsub server at this time. (the extra
deps needed to use "uv" to run the tests)